requirements.txt
requests==2.32.3
asyncio==3.4.3
asgiref==3.8.1
Django==5.0.2
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.2
django-storages==1.14.3
dj-database-url==2.2.0
psycopg[c]==3.1.18
python-dotenv==1.0.1
langchain==0.2.15
langchain_community==0.2.14
langchain_core==0.2.36
langchain_openai==0.1.23
langchain_text_splitters==0.2.2
beautifulsoup4==4.12.3
azure-core==1.30.2
azure-storage-blob==12.20.0
pillow==9.5.0
gunicorn==20.1.0
django-temporalio==1.1.0
tavily-python==0.5.0
ag2[captainagent]==0.5.0
pymongo==4.10.1
markdownify
arxiv
pymupdf
wikipedia-api
easyocr
python-pptx
openai-whisper
pandas
scipy
sentence-transformers
